refactor: isortを適用, データクラスのdirをschemaに変更
