+++ packages/backend/pyproject.toml
@@ -11,6 +11,7 @@ typer = "^0.6.1"
 boto3 = "^1.24.80"
 postgres = "^4.0"
 SQLAlchemy = "^1.4.41"
+isort = "^5.10.1"
 
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
